Update BaseShadowSystem
Remove first pass from base shadow system and create new class
FirstPassBaseShadowSystem.
